Major work completed
Added a lot to the website, including animated buttons for the four pages that the site will have, a picture and setup for an about me section.

Added some buttons in the footer that link to my github and linkedin pages.

Added a js file for clean transitions between sections.